Match: script to generate debugging input taken in bin/match/
Until now, to generate inputs to the Match program, we always needed two
catalogs and it was annoying to build them by hand every time. So during
the work on k-d trees, a script was written to either build two catalogs
from scratch, or use an existing one and add noise to its coordinates. But
that script was in a temporary directory we had built while working on
Match.

With this commit, that temporary directory has been deleted, however, since
the input generation script was useful for the future, it has been put in
the 'bin/match/' directory in case it may be useful in the future.