Book: corrected placing of the --asis option of astfits
Until now, in the Gnuastro book, the '--asis' option of the Fits program
was mistakenly placed as a sub-item of the '--write' option
features. However, it is an independent operation!

With this commit, it has been brought out of the sub-items of '--write' and
placed under it to avoid confusing the readers.