Crop: --widthinpix option for width to be read as pixels in WCS mode
Until now, in Crop's WCS mode, the value to '--width' had to be in the same
units as the WCS (degrees for RA, Dec). However, this was inconvenient when
you had coordinates in RA/Dec, but wanted crops of a certain width (in
pixels): you would have to get the pixel scale with 'astfits' for example,
then convert your desired pixel-width to degrees.

With this commit, Crop now has a new '--widthinpix' option (it is just an
on-off option: meaning that it doesn't take a value). When this option is
called, and Crop is in WCS mode, the value that the user gives to '--width'
will be interpreted as pixels, not degrees.

This new feature has also been included in the warning message that is
printed when the crop size is larger than 50000 pixels in WCS mode. This
usually happens when the user forgets that in WCS mode by default '--width'
is in units of degrees, not pixels, and thus gives values like
'--width=500'!

This feature was suggested by Jesús Varela.