PSF stamp: new option to set a mask sky pixel value: maskskyval
Until this commit, when using a mask (labeled image) this script assumed
that sky background values (pixels that don't have to be masked) were equal
to 0. The reason is that we used to treat mask images from Gnuastro. For
Gnuastro (NoiseChisel and Segment outputs) the sky is considered as 0 in
the detection and segmentation outputs. However, there may be other software
that don't follow this convention. In that situation it is good to have an
option to specify what is the sky background pixel value on the mask image.

With this commit, this new option has been added: --maskskyval or -v. Now
the user is able to specify the sky background value of the mask image. A
bug has been also corrected. Until now, when the cropped core image (those
that is used to detect the central object and not mask it) had more sky
background pixels than the labeled object, the central object was
considered to be the sky (just because there are more pixels!). This is a
bug and it has been corrected by changing the sky background pixel values
to nan values before computing the pixel value of the central object.

This bug was reported by Pablo Manuel Sanchez Alarcon.