Book: corrected example of --catrowfile
In the previous commit, an example command was given for the usage of
'--catrowfile' (to add rows from different tables into one table). But the
example didn't exactly correspond to the description in the text (it didn't
have any reference to the RA, DEC and MAGNITUDE columns that were
mentioned).

With this commit, the command has been corrected to include those columns.