Match: added tests, completed docs of new k-d tree based matching
Until now I was busy implementing, testing, and imptoving the various
phases of the k-d tree based matching. Finally, things seem to be in place
properly so it was necessary to add tests in 'make check' and complete the
uncomplete documentation of this new feature (with the new "Matching
algorithms" section in the book).

With this commit, some tests were added, and the functions were cleaned and
made more clear, while fixing some problems that occurred in the previous
implementation (for small tables, edge-effects of the binning algorithm
were causing problems, and not letting matches getting found). Also some of
the new features have now been explained in 'NEWS'.