Book: For selecting stars some mandatory options did not mention
Until now, in book for selecting stars in different range of magnitude,
from '--min' and '--max' option was used. And some options such as
'--minaxisratio', '--mindistdeg' and '--matchaperturedeg' was defined as
optional options.

Now, from '--magnituderange' have been used for selecting stars in
different range of magnitude instead of '--min' and '--max' options. The
'--minaxisratio', '--mindistdeg' and '--matchaperturedeg' options are
mandatory now and in examples of part of selecting stars, these options
have been used.