Book: correct the definition of option --minaxisratio
Until now, in definition of the '--minaxisratio' explained when the user
did not provide this option by default the value of this option was '0.9'
and this option was optional.

Now, the '--minaxisratio' is one of the mandatory options and when the user
does not provide this option the script will be stopped and definition part
has been corrected.