Library (txt.h): correctly counting existing metadata
Until now, we were incorrectly counting the number of actual data in a
plain-text file: the counter was incremented as the plain text file was
being parsed, so by the end of the line, the counter showed one value
larger than the actual number of columns. As a result, when there was more
metadata than actual columns (for example having f '# Column N:'s but only
three actual data columns), the table-parsing would crash in the sanity
checks of later phases.

With this commit, after the line of data is parsed, the counter is
decreased by one, to ensure that it shows the actual number of
columns. This fixed the problem.

This bug was found thanks to Zohreh Ghaffari.

This fixes bug #61490.