NoiseChisel: --checksn includes pseudos that are right on area limit
Until now, when the user ran NoiseChisel with '--checksn' (to see the list
of pseudo-detection S/Ns that are used to determine the final true/false
S/N threshold), the output table wouldn't contain the those
pseudo-detections that had exactly the area given to '--snminarea'! Only
pseudos that were larger than that area were included in the table.

With this commit, the problematic part of the code has been found and
fixed. In short, instead of using '>=' (greater equal), I was using '>'
(greater than)! In the process, the debugging part of the code (that is
commented) was also updated to be more generically usable.

This bug was found by Raúl Infante-Sainz.

This fixes bug #61493.