Option parsing: all coordinate columns accept sexagesimal
Until now, the coordinate parsing options of all the programs would only
accept RA and Dec in degrees (a single number). However, in many contexts,
we are provided with sexagesimal RAs and Decs and having to convert them to
degrees before calling the programs was annoying.

With this commit, the low-level option parsing options have been edited to
attempt to read the given value as a sexagesimal coordinate also (in case
they can't be read as a single number).

Besides convenience on the command-line, with this feature, users don't
have to worry about how DS9 records the polygon vertices in its region file
(by default DS9 saves values in sexagesimal, and this would cause a crash).

This was initially suggested by Alessandro Ederoclite for the '--center'
option of Crop, and also related to the DS9 region file parsing bug that
was reported by Juan Antonio Fernández Ontiveros and Alejandro Lumbreras
Calle.

This fixes bug #61468.