Book: corrected typos in some file names of the first tutorial
Until now, in the first tutorial on simulating a mock galaxy, in a few
cases, the file names mentioned, didn't match the file names that the user
got after running the command. This was because of a recent feature added
to MakeProfiles (to add '_profiles.fits' suffix to its output when none was
given).

With this commit, the file names have been corrected in the text of the
tutorial.

These typos were reported by Zahra Hosseini.