PSF scripts: getting proper NAXIS values when data is compressed as .fz
Until this commit, if some of the input images were compressed as .fz
files, the computation of the dimension of the images were not properly
done. This is because it was reading the keywords NAXIS1 and NAXIS2.
However, when the data is compressed, these keywords are not the original
image dimensions but the dimensions of the table (data compressed).

With this commit, this bug has been corrected. Now, the scripts read the
keywords: ZNAXIS1, ZNAXIS2, NAXIS1, and NAXIS2. If ZNAXIS keyword values
exist, then use them as the original image dimensions. Otherwise, the size
will be computed from the normal NAXIS and NAXIS2 keywords.