Book: add the new section 'skewness of the image' to the book.
Until now: in image surface brightness limit section the quantile of the
mean was added instaed of standard definiton of skweness. But this section
is too long, we could not explain about the difference of between quantile
of the mean and standard definition of skweness.

Now, new section 'skweness of the image' is added to the book and in this
section about the difference between quantile of the mean and standard
definition of the skweness are explained in more details.