Library (checkset): add timestamp function
Until now, when it was necessary to create temporary debugging files during
the development, there was no generic/easy way to generate their names in a
systematic way. As a result, developers would be discouraged from making
such files (that would ultimately slow the development cycle).

With this commit, a new function called 'al_checkset_timestamp' has been
added to Gnuastro's internal libray. This new function receives a filename
(or any string), checks current system time, adds a YYYY-MM-DD prefix, a
HH-MM-SS suffix, and replaces the file extension (if any) with given
extension.

The date prefix and time suffix are added in this order so that sorting
same files within a working day will be easier. Same files will be stacked
on top of each other regardless of their modification time, but based on
their modification date!