Fits: CHECKSUM automatically updated if any keywords change
Until now, when any of the keywords were changed with the relevant options
of the Fits program and the HDU already had a 'CHECKSUM' keyword, the
change would invalidate the 'CHECKSUM'. The only way to avoid this was to
call '--write=checksum' along with other modifications (as described under
'--write' in the manual). For example the command below to delete the 'GSL'
keyword.

    astfits a.fits --delete=GSL --write=checksum

However, this was only useful for those keyword modifying operators that
have less precedence than '--write' (precedence is also shown as a numbered
list in the manual). As a result, it was is necessary for Gnuastro's Fits
program to check if the file has a 'CHECKSUM' keyword or not. If 'CHECKSUM'
is present, then upon any change in the keywords (Fits doesn't change the
data!), the 'CHECKSUM' keyword should also be updated in the end.

With this commit, this task has been done. In the process, I also noticed
that we don't have a simple function to check the existance of a keyword,
so a new function called 'gal_fits_key_exists_fptr' has also been
added. Because of this change, the library so-name version was also
incremented (it was still the same version as the published Gnuastro 0.16).

This task was suggested by Tamara Civera Lorenzo.