Webpage: corrected Peertube links in Videos section
Until now, on the main webpage we were linking to 'peertube.tv' URLs
because this was the webpage that they were initially uploaded to. However,
apparently Peertube has moved the videos to 'peertube.stream' URLs! So when
a viewer would click on the Gnuastro pager's URLs, Peetube would ask to
redirect them.

With this commit, the URLs have been corrected to take the viewers directly
to the actual video URLs without any redirection.

This issue was reported by Sergio Chueca Urzay.