Webpage: corrected top version URL to 0.16.1
Until the previously commit, I had corrected the Gnuastro tarball URL on
the main webpage to be 0.16.1 (due to bug #61329). But I had forgot to
commit that change in the Git history!

With this commit, this issue has been corrected and the English and French
webpages within the version-controlled source also link to the version
0.16.1 version and URL.