Configure script: Gnulib checks ignore high-level check results
Until now, Gnulib was simply inheriting its build environment after
possible modifications by the higher-level (mostly astronomical software)
checks. This would cause conflicts on some systems where the low-level
software have the same names as the high-level ones (like WCSLIB, which
shares the name 'libwcs' with the system's "wide character string"
library!).

In Gnuastro, the configure-time checks of the GNU Portability Library
(Gnulib) are placed after the checks for the astronoimcal
dependencies. This is because Gnulib is very low-level and has MANY checks!
But Gnulib is designed to find ways to resolve those checks, so they will
never crash. But if a user is missing an astronomical dependency, its best
that the script crashes earlier.

With this commit, the problem has been addressed by temporarily removing
the edits to the environment before Gnulib, so it can check the low-level
system settings without trouble.

This bug was reported by Martin Guerrero Roncel.