Arithmetic: checking existing output name, also when no .fits suffix
Until now, when an output name was given, but the input file name didn't
have a '.fits' suffix, Arithmetic would not delete the output file (to
write from scratch), it would simply add a new extension to the already
existing file.

With this commit, a check has been added to check the contents of a file
also and remove the output file if it already existed.

This fixes bug #61584.