Library (fits.h): using fits_is_reentrant for parallel reading of tables
Until now, the new parallelized table-reading feature of
'gal_fits_tab_read' was simply using the
'GAL_CONFIG_HAVE_FITS_IS_REENTRANT' macro as a reference to see if
CFITSIO's multi-threaded file reading is available or not. However, this
was wrong! That macro was only in place to make sure that the function
providing this test result is available or not!

With this commit, we are now using this macro and the CFITSIO function
properly. Also, a message is now printed if the user's CFITSIO version is
not up to date (and doesn't have the function to see if it allows parallel
access to a file).

This issue was found by Raul Infante-Sainz.