Book: typo fixed in section Distance on a 2D curved space
Until now, while deriving the change in distance on a flat 2D space, we had
forgot to add the power-of-two on the left-side of the equal sign!

With this commit, the power-of-two has been added, and a minor edit has
also been made in the paragraph above to be more clear.

This typo was found by Zohreh Ghaffari.