Table: new --noblankend option to allow removing blanks from input
Until now, the '--noblank' option would only apply to the output table (for
example after doing column arithmetic or adding columns from other
tables). However, logically, it is very similar to the '--range' or
'--equal' options (that operate on the input table before column arithmetic
or addign columns). As a result, this differing behavior could cause
confusion for the users!

With this commit, the '--noblank' operator now operates on the input table
(and doesn't necessarily have to be in the output table!). To keep the old
functionality, the new '--noblankend' option has been added that has the
same behavior as the old '--noblank', except that it can be called multiple
times, not just once with all the columns.

This issue was reported by Samane Raji.