PSF scripts: fixed two FAILS during make checks
Until this commit, there were two PSF scripts that failed during the make
check steps. For the 'psf-create-junction' script the problem was that the
variable of the inner image name was not the correct one: it was 'psf' but
should be 'core'. For the 'psf-create-select-stars' the problem was that
when using the option '--range' in Table, the separator should be commas
',', however, semicolons ':' were used instead.

With this commit, these bugs have been corrected. Now the make checks run
without any problem for this two scripts.