book: update gal_polygon_clip definition
Until now, the gal_polygon_clip had a small definition in the
manual. However, it was defined pretty well in the lib/polygon.c file.

With this commit, I have relocated the detailed definition to the Gnuastro
manual. Now, developers can read more about it in the manual itself,
without the need to visit the lib/polygon.c file for a better understanding
of gal_polygon_clip innerworkings.