Library (arithmetic.h): NaN used in multioperands with float output
Until now, when the inputs to the 'mean' operator were integer types, and
some of the output pixels weren't covered by any of the inputs, instead of
a NaN value, those pixels contained large negative values (for example
-2.14748e+09 when the inputs are 32-bit integers).

When the inputs to the 'mean' operator don't contain any data in a pixel,
the output pixel should have a NaN value (the output of 'mean' is always a
floating point numerical data type, irrespective of the input's data
type). The same applies to the 'sum', 'std', 'median' and all 'sigclip-*'
operators.

The issue was caused because we were simply writing the input's blank value
in these pixels of the output (hence a value of ~2 billion for a signed
integer). But this was wrong, because the output of these operators doesn't
have the same type as the input.

With this commit, for these operators instead of using the input's blank
value, we are simply writing a NaN when there is no data in that pixel.

This bug was reported by Zohre Ghaffari.

This fixes bug #61698.