Library (arithmetic.h): using 0 for sigclip-number instead of blank
Until now (in the previous commit), we simply used NaN for all the
'sigclip-*' operators of the arithmetic library. However, for the
'sigclip-number' operator, blank is meaningless! Its just a counter and a
counter is 0 for the case that there is no data.

With this commit, when 'sigclip-number' is used and there is no data,
instead of a NaN, we write 0.