Library (options.h): accounting for non-sexagesimal colons in values
Until now, when 'gal_options_parse_list_of_numbers' was given a string like
'4:5' (extracted for example from Table's '--range=NAME,4:5') it would
abort, complaining that '4:5' can't be parsed as a number. This is the
result of a recent change in this function to allow reading sexagesimal
coordinates for example '01:23:45.67'.

With this commit, the issue was fixed by avoiding the stopping of the
program when the first un-parsed character was a ':'.

It wasn't registered as a bug because it was due to a recent addition
(after version 0.16).

This issue was reported by Sepideh Eskandarlou.