Book: highlighted usage of comma for --range in asttable
Until now, in the description of Table's '--range' option there was no
explanation that a comma (',') can also separate the values.

With this commit, a sentence has been added to let users know that the
'--range' option in Table, can separate the value with ':' (like
'--range=magnitude,3:5') and with ',' (like this '--range=magnitude,3,5').