developer-build: check added to see if bootstrapping has been done
Until now, if the 'developer-build' script was run without bootstrapping,
it would simply abort with a './configure: No such file or directory'
error. The configure script is created by the bootstrapping process, but
that is not directly obvious.

With this commit, a check has been added at the start of the
'developer-build' script to see if 'configure' exists or not. If it doesn't
exist, the script stops with a descriptive message, showing how to fix the
problem (by bootstrapping).

This issue was reported by Sepideh Eskandarlou and Zahra Hosseini.