Book: improved photon-counting noise description with analogy
Until now, there wasn't any description for adding an appropriate value of
background when we used in '--background' options of 'astmkprof'. The
connection of the background value to the photon-counting noise was also
not very well elaborated in that section!

With this commit, we bring up an analogy in the "Photon counting noise"
subsection of the "Noise basics" section: comparing the background value to
the depth of a sea (such that the height of waves/noise corresponds to the
depth of the sea). We also explain the importance of background value in
understanding this source of noise. For this aim, we further explain the
relation between the background value and the free parameters of the
noise's Gaussian distribution.

This was done with the help of Sepideh Eskandarlou.

The text was slightly edited by Mohammad Akhlaghi upon merging into the
'master' branch of Gnuastro.