Book: commands to check bootstrapping dependencies added
Until now, the Bootstrapping dependencies section of the book didn't have
any way for people to make sure that they have actually setup the
tool. Also, the list of dependencies was missing GNU Texinfo!

With this commit, GNU Texinfo has been added to the list and an example
command has been added to help check if the necessary tool exists.

This was suggested by Sepideh Eskandarlou and Zahra Hosseini.