Copyrights: ending year updated to 2022
Until now, the ending years in the copyright statements of Gnuastro's
source were 2021. But 2022 has started from today, so it was time for an
update!

With this commit, all the ending copyright years in all the Gnuastro source
files have been changed to 2022.

Since the first commit of Gnuastro (09ac8974fa51) was on the 1st of January
in 2015, Gnuastro is now officially 7 years old! Unofficially (before that
first commit), its one year older!