Book: edits in the PSF scripts
Until now, I hadn't had a look at the newly added section describing the
extended PSF creation scripts, or actually tried them out!

With this commit, some edits were made in the book to help in
readability. In particular, the long paragraphs were shortened and in the
introduction, instead of simply using the script names, we now have a
'@ref' to their respective section (that also includes their names).

I also noticed a small issue in the 'psf-create-select-stars' script that
was due to the new change in behavior for '--noblank' (which now operates
before column arithmetic). So it was changed to '--noblankend'.

In the tutorial, I also removed the numbers as variables that were only
used once. Indeed, for a script it is necessary to have these as variables,
but for a tutorial, they just furstrate the reader who is tyring to make
sense of everything. Its easier to actually write the value in the text,
give context, and describe why that choice was made, and use that value in
the command.

Also, I have re-formatted the outputs of the star stamps phase to be in a
separate directory and have simple names instead of very long names
containing the RA and Dec of stars! Such names are hard to read, write,
find or understand!

However, I stopped doing the tutorial because of the star at
(201.94042,47.75882): it has a very bright neighbor! The select-stars
script should have rejected this star.

Sepideh, please correct this in the script (or how its called in the
tutorial) and let me know so I look into the rest of the work in this
branch.