Table: corrected wrong short option for --catrowhdu
Until now, the short-option name for the recently added '--catrowhdu' was
'-H', which was the same short-option name for '--head'! As a result, all
values given to '--head' were being passed to '--catrowhdu'!

With this commit, '--catrowhdu' has been given a new short-option name of
'-X', and this solved the problem. While doing this check, I noticed that
'--catrowhdu' didn't have any default value in the Table configuration
file.

Furthermore, I noticed that there is no sanity check to avoid calling both
'--head' and '--tail' at the same time! So a check has been added to alert
the user when it (mistakenly!) happens.

This issue was reported by Manuel Sánchez-Benavente.