PSF select-stars: remove the stars that have stars.
Until now, the circular stars with good parallax and less than 9 neighbor
in the range of magnitude between brighter to fainter from the gaia or
catlog have been selected. But some of the main circular star with good
parallax and less than 9 neighbor have a neighbor that its brightness are
so close to the main star, but it did not remove in all above conditions
(parallax circular and neighborhood).

With this commit, for tackling this problem at first a limit for selecting
the faint and bright stars hase been set, that select stars in the wider
range of magnitude compare to the brighter and fainter range. For setting
these limit for bright stars a fix value have benn considered (-10) because
we should consider all brightest stars around the main star, for faint
stars this limit is optional and the user can determined this limit with
'--faintmagsiff' to select the stars that are manipult brighter than
fainter magnitude. Now we have a wide range of stars for checking the
number of the neighbors in the range of magnitude between brighter to
fainter. Finally when the number of the main stars has been checked just
stars without any neighbor will be selected.