PSF select-stars: change the name of the two varibales.
Until now, the two variable 'minmag' and 'maxmag' has been used in script
for selecting the stars in the range of magnitude between minimumm to
maximumm but these two name will be confused the user and she/he do not
know the 'minmag' option has been used for the bright or faint star.

With this commit, for the clarity the name of two these variable has been
changed from 'minmag' to 'brightmag' and from 'maxmag' to the 'faintmag'.