PSF select-stars: add new option for selecting more bright stars
Until now, for select the bright stars in wider range than user has been
determined, the constant value had been considered. And then based on this
constant value the wide bright stars had been selected and find the number
of neighbors around the main star.

With this commit, the new option '--brightmag' has been added, that
selecting the bright stars in the wider range to be optional.