PSF select-stars: give the default value to the two variables
Until now, for the two varialbe '--faintmagdiff', '--brightmag' had not
default value determined and when the user did not determine these values
the script will be interrupt and wanted the user determine these valuse.

With this commit, the default value has been defined for two these
variables '--faintmagdiff', '--brightmag' and if the user does not
determind the value of these parameters the default value has been
considered.