Configuration: no modification in LDFLAGS for library checks
Until now (since commit 1ac084b4), the LDFLAGS library was being
temporarily modified before checking for high-level features of the
available libraries (for example if 'pthreads' has the barrier feature, or
if WCSLIB has a version function and etc). On some systems (for example
Ubuntu 20.04), this would cause crash during the 'make' step (and in
particular because of the conflicting definitions of the pthread's barrier
feature!).

This particular change in LDFLAGS was not the cause of the problem behind
commit 1ac084b4 (that was to do with 'LIBS' and Gnulib!). It was just a
change that didn't cause a conflict on my system, so I thought it should be
fine (given the similar change that I did to fix the Gnulib problem).

With this commit, that particular part of temporarily changing LDFLAGS in
commit 1ac084b4 is now removed from the configure time checks and this
fixed the problem.

This issue was reported by Raul Infante-Sainz.