PSF tutorial: running of Segment on saturated images added
Until now, the tutorial that demonstrated the usage of the PSF scripts
didn't give any solution to using Segment to identify clumps for masking of
the background sources.

With this commit, the tutorial has been broken into separate sub-sections
and a new sub-section has been added for describing how to have a unified
clump even though there are saturated regions in the centers of the stars.

Also, I noticed that some of the options in the 'psf-create-select-stars'
didn't have any default values. They have been were given default values,
to make it easier for a user to run them (and not be forced to give many
options on every run).