Book: Edits to the tutorial on creating the extended PSF
Until now, there was no top-level introduction in the tutorial on creating
an extended PSF and the various files generated in the various phases were
being written in the running directory.

With this commit, both issues have been addressed until the part where we
make normalized stamps of the selected out stars.