Book: PSF construction tutorial, completed until the outer part
Until now, the usage of 'astscript-psf-create-mask-stamp' script didn't use
Segment's output, and when a '--mask' was given, it would only use a single
mask: not benefiting from the fact that Segment provides both clumps and
objects. Since Segment's clumps and objects are ideal for such scenarios
and this script is built to run within Gnuastro, and to avoid complicating
the code, its much easier for the user to simply give the output of Segment
(when saturated pixels have been properly dealt with; as described in the
first part of the tutorial).

With this commit, the following changes were made to
'astscript-psf-create-mask-stamp' (these are not yet documented in the
respective "Invoking ..." section of the book):

 --segment: the new name for '--mask', and it now assumes that the file
   given is the output of Segment.

 --maskhdu: removed because we now assume the output of Segment.

 --saturation: removed because we now describe how to properly deal with
   saturated pixels in the respective tutorial.

 --corewidth: default value set to '5,5'. This is because we now explain
   how to properly deal with saturated pixels, so using a very small region
   will give a more robust estimate of the clump and object IDs for the
   user's given coordinate.

 - I noticed that '--mask' was given two times in the output of '--help'!

In relation to these changes, the tutorial was updated until the part where
the outer PSF is created (using three stars with magnitude range 6 to
8). Some new sections have also been added that need to be filled before we
can finish this branch.