Book: PSF tutorial now with instruction to build first inner region
Until now, we had just stopped the tutorial at the stage of building the
outer PSF. As part of the on-going effort to conclude with a full PSF,
steps needed to be added for building the inner parts also.

With this commit, the first inner region stack has been created. Also, the
image ID of J-PLUS DR2 data was added to the produced files, to encourage
people to use other tiles also (and build a better PSF with more stars).