Library (type.h): properly reading many digit numbers as double
Until now, when a many-digit "small" number (like '330624.3918430004') is
given to 'gal_type_string_to_number', it is mistakenly read as a 32-bit
float number, not 64-bit. This causes problems like the example below (when
reading FITS keywords that have values in a similar format). This
'img.fits' can be any of the images that is initially downloaded in the
general tutorial:

    $ astfits img.fits -h0 --keyvalue=TEXPTIME
    img.fits 330624

    $ astfits img.fits -h0 | grep TEXPTIME
    TEXPTIME=    330624.3918430004

As you see, while the value is '330624.3918430004', the Fits program only
prints the first 6 characters! The root cause of this problem is that
'gal_type_string_to_number' is wrongly identifying this as a 32-bit
floating point number!

The cause of the problem was found in the wrong limit checks of
'gal_type_string_to_number': previously if the number was smaller than the
maximum float32 or larger than the minimum float32, the number of its
digits was effectively ignored (there was an "or").

With this commit, the condition check was fixed: we identify a number as
float64 when the number of its meaningful digits is larger than 'FLT_DIG',
or its absolute value is larger than the maximum float32 or smaller than
the minimum float32 (more precision). This fixed the problem.

This bug was first reported by Zohre Ghaffari and Pedram Ashofteh Ardakani
helped in finding the solution.

This fixes bug #61940.