Library (ds9.c): account for DS9 options after polygon definition
Until now, when the 'ds9.c' library was reading for the definition of a
polygon in the respective line, it expected that it ends with a
')'. Otherwise, it would abort the program, and complain about the
formatting.

This raised problems when the region file contained any user-specified
non-default metadata (like colors or text) following the line starting with
'polygon(...)' in the same line (for example '# color=red').

With this commit, instead of simply checking the last character of the line
(and aborting if its not a ')'), we actually parse the line and stop on the
first ')'. If we confront a '#' before that, a status variable will be set
and the parsing will stopt.