PSF scripts: keeping temporary directory in radial profile calls
Until this commit, when constructing the PSF, the temporary directory was
removed as soon as the radial-profile script finished. This is simply
because the temporary directory for the PSF scripts and the temporary
directory for the radial profile script was the same. This is not good
since the directory is removed in the middle of the execution.  Step by
step, what happened was:

 - PSF scripts generated a temporary directory (tmpdir).
 - At some point, the PSF scripts called the radial profile script.
 - The radial profile script removed the temporary directory (tmpdir).
 - PSF scripts crashed as a consequence of the directory removal.

With this commit, this bug has been corrected by keeping the temporary
files when calling the radial profile script (within the PSF scripts). By
doing this, the temporary directory is kept until the end of the execution
of the PSf script and the scripts don't crash.