Developer script: add --prefix option
Until now, the 'developer-build' script would set up the 'configure' file
using the default '/usr/local' prefix. This would install all
architecture-independent files there, i.e. PREFIX. To install on this path,
the user needs write permissions, which is given by sudo. However,
sometimes the users want to install gnuastro somewhere without the need of
root permissions.

With this commit, the '--prefix' or '-o' option is added to the developer
script. This will be given to the 'configure' file as value of '--prefix'
option. Also, in case the user has asked for installation with the '-i'
option, the script checks for writing permission. It would install if the
user has writing permission, if not, it uses 'sudo'.