Arithmetic: checking/deleting output file (if needed) in any case
Until now, when the command-line arguments (tokens) to the reverse polish
notation didn't include a FITS file (for example with the 'makenew'
operator), Arithmetic wouldn't make sure it is removed (if the user hasn't
called '--dontdelete') before it starts writing its output. As a result,
when the output file already existed, the output dataset would be written
in a new extension of the output file.

But this is not intuitive in Gnuastro: all programs (including Arithmetic;
when a FITS file exists in the arguments) will check the existance of the
output and delete it.

With this commit, the source of this bug has been found and fixed! We were
checking for an output file name's existance only while parsing the tokens
and when we found a FITS file! This was done because we needed a base-name
for the output file if no '--output' was given. However, as a result, we
were implicitly assuming that 'makenew' is either called with another FITS
file, or without '--output'!

The check has now become generic, so we don't check for the existance of
the output file while parsing the tokens. We just set the basename and
unify the checking for the output file afterwards (to be useful in all
cases).

This fixes bug #62008.