Table: --equal, --notequal and --colmetadata can contain comma
Until now, the column values/information given to these options couldn't
contain a comma. This was because a comma was used as a delimiter between
the values!

With this commit, by adding a '\' before a comma, users can use a comma in
any of the values (not names!) of the options.

This task was suggested by Zohreh Ghaffari.