Arithmetic: only a 1D single number will be printed on terminal
Until now, when Arithmetic ends up with a single number, it will simply
print the number on the command-line. This is good for a 1 dimensional,
single-valued dataset; but when the dataset can have multiple dimensions,
it is annoying. For example the command below will fail to create a 2D
image with a single element:

    astarithmetic 1 1 2 makenew

With this commit, the following two conditions have been added for printing
the value on the command-line: the dataset has to be one-dimensional, and
the '--onedasimage' should be called.