Arithmetic: new repeat operator
Until now, the only way to repeat a certain dataset on the stack was to
manually repeat its name on the command line. This is very inconvenient and
could even cause problems on systems that have a limited buffer for the
command-line (when thousands of copies are necessary!).

With this commit, a new 'repeat' operator has been added to Arithmetic for
this purpose. It is now very easy to make thousands of copies of a dataset
if necessary.