Table: date-to-millisec is a new column arithmetic operator
Until now, when 'date-to-sec' was given a time-string with sub-second
precision, it would add that fractional value to the large integer and save
the result as a 64-bit floating point type.

However, the number of significant digits in the floating point standard is
limited (its in logarithmic scale) and as a result for two different time
strings like below, it would give the same result (even though they are
more than one second apart.

    date                        seconds
    ----                        -------
    2018-05-04T04:16:46.7774    1525410945
    2018-05-04T04:15:25.8326    1525410945

With this commit, the issue has been solved in two steps:

 - 'date-to-sec' operator now only returns 64-bit signed integer numbers
   (no more floating points), and will therefore ignore any sub-second
   precision.

 - 'date-to-millisec' operator has been added to column arithmetic to let
   users count (as an integer: thus preserving every digit properly) the
   number of milli-seconds since the Unix epoch. Users who need sub-second
   precision can use this.

As a result of these changes, the two dates above can now be read in units
of seconds (second column) and milli-seconds with all digits properly
preserved:

    date                        seconds       milli-seconds
    ----                        -------       -------------
    2018-05-04T04:16:46.7774    1525411006    1525411006777
    2018-05-04T04:15:25.8326    1525410925    1525410925832

Notice how the correct seconds of both dates is now different from the
common one that was initially reported (when converting it to 64-bit
floating point).

This bug was reported by Zohreh Ghaffari and Pedram Ashofteh Ardakani.

This fixes bug #61976.